requirements.txt
Flask==3.0.0
spotipy==2.23.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
python-dotenv==1.0.0
requests==2.31.0
